devtools/server/tests/mochitest/test_inspector-insert.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript" src="inspector-helpers.js"></script>
  <script type="application/javascript">
"use strict";

const {DocumentWalker} = require("devtools/server/actors/inspector/document-walker");

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
};

let gWalker = null;
let gInspectee = null;

addTest(function setup() {
  const url = document.getElementById("inspectorContent").href;
  attachURL(url, function(err, client, tab, doc) {
    gInspectee = doc;
    const {InspectorFront} = require("devtools/shared/fronts/inspector");
    const inspectorFront = InspectorFront(client, tab);
    promiseDone(inspectorFront.getWalker().then(walker => {
      ok(walker, "getWalker() should return an actor.");
      gWalker = walker;
    }).then(runNextTest));
  });
});

addAsyncTest(async function testRearrange() {
  const longlist = await gWalker.querySelector(gWalker.rootNode, "#longlist");
  let children = await gWalker.children(longlist);
  const nodeA = children.nodes[0];
  is(nodeA.id, "a", "Got the expected node.");

  // Move nodeA to the end of the list.
  await gWalker.insertBefore(nodeA, longlist, null);
  ok(!gInspectee.querySelector("#a").nextSibling,
     "a should now be at the end of the list.");
  children = await gWalker.children(longlist);
  is(nodeA, children.nodes[children.nodes.length - 1],
     "a should now be the last returned child.");

  // Now move it to the middle of the list.
  const nextNode = children.nodes[13];
  await gWalker.insertBefore(nodeA, longlist, nextNode);
  const sibling =
    new DocumentWalker(gInspectee.querySelector("#a"), window).nextSibling();
  is(sibling, nextNode.rawNode(), "Node should match the expected next node.");
  children = await gWalker.children(longlist);
  is(nodeA, children.nodes[13], "a should be where we expect it.");
  is(nextNode, children.nodes[14], "next node should be where we expect it.");

  runNextTest();
});

addAsyncTest(async function testInsertInvalidInput() {
  const longlist = await gWalker.querySelector(gWalker.rootNode, "#longlist");
  const children = await gWalker.children(longlist);
  const nodeA = children.nodes[0];
  const nextSibling = children.nodes[1];

  // Now move it to the original location and make sure no mutation happens.
  let hasMutated = false;
  const observer = new gInspectee.defaultView.MutationObserver(() => {
    hasMutated = true;
  });
  observer.observe(longlist.rawNode(), {
    childList: true,
  });

  await gWalker.insertBefore(nodeA, longlist, nodeA);
  ok(!hasMutated, "hasn't mutated");
  hasMutated = false;

  await gWalker.insertBefore(nodeA, longlist, nextSibling);
  ok(!hasMutated, "still hasn't mutated after inserting before nextSibling");
  hasMutated = false;

  await gWalker.insertBefore(nodeA, longlist);
  ok(hasMutated, "has mutated after inserting with null sibling");
  hasMutated = false;

  await gWalker.insertBefore(nodeA, longlist);
  ok(!hasMutated, "hasn't mutated after inserting with null sibling again");

  observer.disconnect();
  runNextTest();
});

addTest(function cleanup() {
  gWalker = null;
  gInspectee = null;
  runNextTest();
});

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
